requirements.txt
python-telegram-bot
digitalguide
Pillow
PyYAML
boto3
ctparse
pymongo[srv]
mongoengine